+++ main.js
@@ -7,13 +7,13 @@
 export {
   installCertificateAuthority,
   uninstallCertificateAuthority,
-} from "./src/authority_certificate.js"
+} from "./src/certificate_authority.js"
 
 export {
   createValidityDurationOfXYears,
   createValidityDurationOfXDays,
 } from "./src/validity_duration.js"
 
-export { verifyHostsFile } from "./src/verify_hosts.js"
+export { verifyHostsFile } from "./src/hosts_file_verif.js"
 
-export { requestCertificateForLocalhost } from "./src/localhost_certificate.js"
+export { requestCertificateForLocalhost } from "./src/certificate_for_localhost.js"
