+++ src/certificate_authority.js
